requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.0.2
django-appconf==1.0.6
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
django-simple-history==3.5.0
django-user-accounts==3.3.2
executing==2.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
ipython==8.22.1
jedi==0.19.1
kombu==5.3.5
matplotlib-inline==0.1.6
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.2
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
